L-29: emit full uint256 cleartextAmount on UnwrapFinalized.

+++ contracts/pod/privacy/PrivacyPortal.sol
@@ -881,7 +881,7 @@ contract PrivacyPortal is IPrivacyPortal, IERC7984PortalWrapper, Pausable, Reent
             withdrawal.recipient,
             withdrawalId,
             withdrawalId,
-            uint64(withdrawal.amount)
+            withdrawal.amount
         );
     }
